+++ src/compas/geometry/primitives/polyline.py
@@ -109,7 +109,7 @@ class Polyline(Primitive):
     # ==========================================================================
 
     def __repr__(self):
-        return "Polyline({})".format(", ".join(["{}".format(point) for point in self.points]))
+        return "Polyline([{}])".format(", ".join(["{}".format(point) for point in self.points]))
 
     def __len__(self):
         return len(self.points)
@@ -118,12 +118,15 @@ class Polyline(Primitive):
         return self.points[key]
 
     def __setitem__(self, key, value):
-        self.points[key] = value
+        self.points[key] = Point(*value)
+        self._lines = None
 
     def __iter__(self):
         return iter(self.points)
 
     def __eq__(self, other):
+        if not hasattr(other, '__iter__') or not hasattr(other, '__len__') or len(self) != len(other):
+            return False
         return all(a == b for a, b in zip(self, other))
 
     # ==========================================================================